Avoiding the KeyNotFoundException in JiraItems.cs (#104)
* Avoiding the KeyNotFoundException in JiraItems.cs

The fields [reporter] and [created] were accessed after the look. However, under certain circumstances these fields might be removed within the foreach-loop from the [fields] dictionary and the KeyNotFoundException is then thrown.

By placing these field before the foreach-loop we ensure that the values are always in the dictionary and so avoid the exception.

* Fix KeyNotFoundException

There might be cases when the reporter is not set and thus causing an exception.